Huge updates
- [modify] Data model types' properties to match what are returned
  from the API
- [add] New data model to fill-in what is missing from the translation
  from javascript API to this Swift API
- Test is passed

+++ Sources/ShopifyAPI/Model/Data Model/TaxLine.swift
@@ -6,7 +6,7 @@ import Foundation
 
 public struct TaxLine: Decodable {
   public var price: String?
-  public var rate: Int?
+  public var rate: Double?
   public var title: String?
   
   enum CodingKeys: CodingKey {
